+++ app/assets/javascripts/initializePage.js.erb
@@ -32,7 +32,6 @@ function callInitalizers(){
   initNotifications();
   initializeSplitTestTracking();
   initializeStylesheetAppend();
-  initializeAnalytics();
   initializeCommentDropdown();
   initializeSettings();
   initializeFooterMod();
