+++ app/controllers/dashboards_controller.rb
@@ -16,10 +16,15 @@ class DashboardsController < ApplicationController
     elsif params[:which] == "user_followers"
       @follows = Follow.where(followable_id: @user.id, followable_type: "User").
         includes(:follower).order("created_at DESC").limit(80)
+    elsif params[:which] == "organization_user_followers"
+      @follows = Follow.where(followable_id: @user.organization_id, followable_type: "Organization").
+        includes(:follower).order("created_at DESC").limit(80)
     elsif @user&.organization && @user&.org_admin && params[:which] == "organization"
       @articles = @user.organization.articles.order("created_at DESC").decorate
     elsif @user
       @articles = @user.articles.order("created_at DESC").decorate
     end
+    # Updates analytics in background if appropriate:
+    ArticleAnalyticsFetcher.new.delay.update_analytics(current_user.id) if @articles
   end
 end
