+++ app/views/dashboards/_analytics.html.erb
@@ -1,0 +1,18 @@
+<% num_views = @articles.sum(&:page_views_count) %>
+
+<div class="dashboard-analytics-header-wrapper">
+  <div class="dashboard-analytics-header">
+    <span><%= number_with_delimiter(@articles.sum(&:positive_reactions_count), delimeter: ",") %></span>
+    <div class="dashboard-analytics-sub-indicator">
+      <img src="<%= asset_path "emoji/emoji-one-heart.png" %>" /><img src="<%= asset_path "emoji/emoji-one-unicorn.png" %>" /><img src="<%= asset_path "emoji/emoji-one-bookmark.png" %>" />
+      Total Post Reactions
+    </div>
+  </div>
+  <div class="dashboard-analytics-header" id="dashboard-analytics-header">
+    <span><%= num_views > 500 ? number_with_delimiter(num_views, delimeter: ",") : "under 500" %></span>
+    <div class="dashboard-analytics-sub-indicator">
+      <img id="dashboard-analytics-image" src="<%= asset_path "emoji/apple-eyes.png" %>" />
+      Total Post Views
+    </div>
+  </div>
+</div>
