+++ config/routes.rb
@@ -107,7 +107,6 @@ Rails.application.routes.draw do
   resources :blocks
   resources :notifications, only: [:index]
   resources :tags, only: [:index]
-  resources :analytics, only: [:index]
   resources :stripe_subscriptions, only: %i[create update destroy]
   resources :stripe_active_cards, only: %i[create update destroy]
   resources :live_articles, only: [:index]
@@ -237,7 +236,7 @@ Rails.application.routes.draw do
   get "/dashboard" => "dashboards#show"
   get "/dashboard/:which" => "dashboards#show",
       constraints: {
-        which: /organization|user_followers|following_users|reading/
+        which: /organization|organization_user_followers|user_followers|following_users|reading/
       }
   get "/dashboard/:username" => "dashboards#show"
